refactor(menu): 메뉴 타입 상수화
