+++ src/main/java/order/model/menu/Menu.java
@@ -1,4 +1,4 @@
-package order.model;
+package order.model.menu;
 
 import order.error.ErrorMessage;
 
@@ -6,19 +6,19 @@ import java.util.Arrays;
 
 public enum Menu {
 
-    PIZZA("피자", 25000, "main"),
-    HAMBURGER("햄버거", 10000, "main"),
-    CHICKEN("치킨", 23000, "main"),
+    PIZZA           ("피자", 25000, MenuType.MAIN),
+    HAMBURGER       ("햄버거", 10000, MenuType.MAIN),
+    CHICKEN         ("치킨", 23000, MenuType.MAIN),
 
-    FRENCH_FRIES("감자튀김", 5000, "side"),
-    CHEESE_STICK("치즈스틱", 7000, "side"),
-    SALAD("샐러드", 8000, "side"),
+    FRENCH_FRIES    ("감자튀김", 5000, MenuType.SIDE),
+    CHEESE_STICK    ("치즈스틱", 7000, MenuType.SIDE),
+    SALAD           ("샐러드", 8000, MenuType.SIDE),
 
-    COLA("콜라", 2000, "drink"),
-    ZERO_COLA("제로 콜라", 2500, "drink"),
-    ORANGE_JUICE("오렌지 주스", 3000, "drink"),
+    COLA            ("콜라", 2000, MenuType.DRINK),
+    ZERO_COLA       ("제로 콜라", 2500, MenuType.DRINK),
+    ORANGE_JUICE    ("오렌지 주스", 3000, MenuType.DRINK),
 
-    SERVICE_DUMPLING("서비스 만두", 0, "service");
+    SERVICE_DUMPLING("서비스 만두", 0, MenuType.SERVICE);
 
     private final String name;
     private final long price;
@@ -42,15 +42,15 @@ public enum Menu {
     }
 
     public boolean isDrink() {
-        return type.equals("drink");
+        return type.equals(MenuType.DRINK);
     }
 
     public boolean isMain() {
-        return type.equals("main");
+        return type.equals(MenuType.MAIN);
     }
 
     public boolean isService() {
-        return type.equals("service");
+        return type.equals(MenuType.SERVICE);
     }
 
     public String getName() {
