+++ src/main/java/order/model/menu/MenuType.java
@@ -1,0 +1,9 @@
+package order.model.menu;
+
+public final class MenuType {
+
+    public static final String MAIN = "main";
+    public static final String SIDE = "side";
+    public static final String DRINK = "drink";
+    public static final String SERVICE = "service";
+}
